@@ -1,65 +1,69 @@
 <template>
-  <div class="x-head" >
-    <img @click="toMain" style="cursor: pointer;height:60px;margin-left: 10px ; margin-top:0.5%" src="headerLogo.png">
+  <div class="x-head">
+    <el-image @click="toMain" style="cursor: pointer;height:60px;margin-left: 10px ; margin-top:0.5%" :src="headerLogo"/>
     <div style="margin:0 auto"></div>
-    <div style="line-height:6vh;margin:0 4vh 0 0" >
+    <div class="place-content-center" style="line-height:6vh;margin:0 4vh 0 0">
       <el-dropdown>
-          <div class="header-right">
-            <span style="color:white">{{this.mail}}</span>
-          </div>
-          <el-dropdown-menu slot="dropdown">
-            <el-dropdown-item icon="el-icon-edit" >
-              <span @click="changePwd()" style="cursor: pointer;">修改密码</span>
+        <span style="color:white">{{ this.mail }}</span>
+        <template #dropdown>
+          <el-dropdown-menu>
+            <el-dropdown-item>
+              <el-icon class="el-icon-edit"></el-icon>
+              <span @click="changePwd" style="cursor: pointer;">修改密码</span>
             </el-dropdown-item>
-            <el-dropdown-item icon="el-icon-edit" >
-              <span @click="changeInfo()" style="cursor: pointer;">个人信息</span>
+            <el-dropdown-item>
+              <el-icon class="el-icon-edit"></el-icon>
+              <span @click="changeInfo" style="cursor: pointer;">个人信息</span>
             </el-dropdown-item>
-            
-            <el-dropdown-item icon="el-icon-s-fold" >
-              <span @click="logout()" style="cursor: pointer;">退出登录</span>
+            <el-dropdown-item>
+              <el-icon class="el-icon-s-fold"></el-icon>
+              <span @click="logout" style="cursor: pointer;">退出登录</span>
             </el-dropdown-item>
           </el-dropdown-menu>
-        </el-dropdown>
-      </div>
+        </template>
+      </el-dropdown>
+    </div>
     <div><br /></div>
   </div>
 </template>
 
 <script>
 import axios from "axios";
+import headerLogo from "@/assets/headerLogo.png";
 export default {
-  name:'headbar',
-  data(){
-    return{
-      mail: localStorage.getItem('Mail')
+  name: 'headbar',
+  data() {
+    return {
+      mail: localStorage.getItem('Mail'),
+      headerLogo: headerLogo,
     }
   },
-  methods:{
-    goToLogin(){
+  methods: {
+    goToLogin() {
       this.$router.push('/studentLogin');
-      localStorage.setItem('Mail',null);
+      localStorage.setItem('Mail', null);
     },
-    logout(){
+    logout() {
       this.$router.push('/studentLogin');
-      localStorage.setItem('Mail',null);
-	  axios({
-		  method:'post',
-		  url: 'user/unLogin',
-		  data: {
-		  	 token:localStorage.getItem('token'),
-		  }
-	  });
-	  localStorage.setItem('token',"wutoken");
+      localStorage.setItem('Mail', null);
+      axios({
+        method: 'post',
+        url: 'user/unLogin',
+        data: {
+          token: localStorage.getItem('token'),
+        }
+      });
+      localStorage.setItem('token', "wutoken");
     },
-    changeInfo(){
+    changeInfo() {
       console.log("Info");
       this.$router.push('/ChangeInfo');
     },
-    changePwd(){
+    changePwd() {
       console.log("Pwd");
       this.$router.push('/ChangePwd');
     },
-    toMain(){
+    toMain() {
       this.$router.push('/borrowRecord');
     }
   }
@@ -69,13 +73,13 @@ export default {
 </script>
 
 <style scoped>
-  .x-head{
-    top:0;
-    left: 0;
-    display: flex;
-    text-align: center;
-    width: 100%;
-    height: 70px;
-    background: linear-gradient(135deg, rgb(88, 184, 228), rgb(14, 86, 168));
-  }
+.x-head {
+  top: 0;
+  left: 0;
+  display: flex;
+  text-align: center;
+  width: 100%;
+  height: 70px;
+  background: linear-gradient(135deg, rgb(88, 184, 228), rgb(14, 86, 168));
+}
 </style>
