@@ -4,23 +4,28 @@
     <div class="contain">
       <div class="big-box" :class="{ active: isLogin }">
         <div class="big-contain" v-if="isLogin">
-          <div class="btitle">忘记密码</div>
+          <div class="btitle">教师登录</div>
           <div class="bform">
-            <span class="count">邮箱 &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;: &nbsp;<input type="text" placeholder="邮箱"
-                v-model="form.email" /></span><br />
-            <span class="count">手机号码 &nbsp;&nbsp; : &nbsp;<input type="text" placeholder="手机号码"
-                v-model="form.phone" /></span><br />
-            <button @click="sendCode">获取验证码<span v-show="!show" class="count"> {{ count }} s</span></button>
-            <br />
-            <span class="count">验证码 &nbsp;&nbsp;&nbsp; &nbsp; : &nbsp;<input type="code" placeholder="验证码"
-                v-model="form.code" /></span><br />
-            <span class="count">设置新密码: &nbsp;<input type="password" placeholder="设置新密码"
-                v-model="form.password" /></span><br />
+            <input type="email" placeholder="邮箱" v-model="form.email" />
+            <span class="errTips" v-if="emailError">* 邮箱填写错误 *</span>
+            <input type="password" placeholder="密码" v-model="form.password" />
+            <span class="errTips" v-if="emailError">* 密码填写错误 *</span>
           </div>
-          <div style="align-items: center; margin-top: 70px">
-            <button class="bbutton" @click="confirmChange">确认修改</button>
-            &nbsp;&nbsp;<button @click="teacherEntry">返回登录</button>
-          </div>
+          <button class="bbutton" @click="login">登录</button>
+          &nbsp;&nbsp;<button @click="teacherForgetPwd">忘记密码</button>
+          <!--  -->
+          <!-- <br>
+          <el-form size="medium">
+            <el-form-item>
+              <el-input v-model="form.email" placeholder="邮箱" type="email"></el-input>
+            </el-form-item>
+            <el-form-item>
+              <el-input v-model="form.password" placeholder="密码" type="password"></el-input>
+            </el-form-item>
+            <el-button type="warning" round @click="login">登录</el-button>
+            <el-button type="warning" round @click="teacherForgetPwd">忘记密码</el-button>
+          </el-form> -->
+          
         </div>
         <div class="big-contain" v-else>
           <div v-if="!stateReg">
@@ -35,7 +40,36 @@
             <br />
             <button class="bbutton" @click="changeRegState">下一步</button>
           </div>
-          <div v-else></div>
+          <div v-else>
+            <div class="btitle">欢迎，{{ this.form.name }}老师！</div>
+            <div class="bform">
+              <input
+                type="text"
+                placeholder="手机号"
+                v-model="form.phone"
+              /><br />
+              <button @click="sendCode">获取验证码<span v-show="!show" class="count"> {{count}} s</span></button>
+              <br />
+              <input
+                type="code"
+                placeholder="验证码"
+                v-model="form.code"
+              /><br />
+              <input
+                type="password"
+                placeholder="密码"
+                v-model="form.password"
+              /><br />
+              <input
+                type="password"
+                placeholder="确认密码"
+                v-model="form.password2"
+              />
+            </div>
+            <button class="bbutton" @click="changeRegState">上一步</button
+            >&nbsp;
+            <button class="bbutton" @click="register">注册</button>
+          </div>
         </div>
       </div>
       <div class="small-box" :class="{ active: isLogin }">
@@ -70,10 +104,13 @@ export default {
       passwordError: false,
       existed: false,
       form: {
-        email: '',
-        password: '',
-        phone: '',
-        code: '',
+        name: "",
+        tid: "",
+        email: "",
+        password: "",
+        password2 :"",
+        code: "",
+        phone: "",
       },
       show: true,
       count: '',
@@ -81,7 +118,7 @@ export default {
     };
   },
   methods: {
-    studentEntry() {
+    studentEntry(){
       this.$router.push('/studentLogin')
     },
     changeType() {
@@ -92,49 +129,50 @@ export default {
       this.form.email = "";
       this.form.password = "";
     },
-    teacherEntry() {
-      this.$router.push('/teacherLogin');
-    },
-    sendCode() {
-      const TIME_COUNT = 60;
-      if (!this.timer) {
-        console.log("已发送"),
-          axios({
-            method: "post",
-            url: 'user/forget/sendCode',
-            data: {
-              acc: this.form.phone,
-            }
-          }).then((res) => {
-            alert("验证码已发送，请查看手机消息。")
-          })
-        this.count = TIME_COUNT;
-        this.show = false;
-        this.timer = setInterval(() => {
-          if (this.count > 0 && this.count <= TIME_COUNT) {
-            this.count--;
-          }
-          else {
-            this.show = true;
-            clearInterval(this.timer);
-            this.timer = null;
-          }
-        }, 1000)
+  teacherForgetPwd(){
+    this.$router.push('/teacherForgetPwd');
+  },
+	sendCode(){
+    const TIME_COUNT = 60;
+     if (!this.timer) {
+       console.log("已发送"),
+		    axios({
+			  method:"post",
+			  url:'user/forget/sendCode',
+			  data:{
+			  	acc:this.form.phone,
+			  }
+        }).then((res)=>{
+        console.log(res.error_code);
+        alert("验证码已发送，请在手机信息中查看。")
+      })
+       this.count = TIME_COUNT;
+       this.show = false;
+       this.timer = setInterval(() => {
+       if (this.count > 0 && this.count <= TIME_COUNT) {
+         this.count--;
+       }
+        else {
+         this.show = true;
+         clearInterval(this.timer);
+         this.timer = null;
+        }
+       }, 1000)
       }
-    },
-    changeRegState() {
-      if (this.form.sid != "" && this.form.name != "" && this.form.college != "") {
-        this.stateReg = !this.stateReg
+	},
+  changeRegState() {
+      if(this.form.sid != "" && this.form.name != "" && this.form.college != ""){
+          this.stateReg = !this.stateReg
       }
-      else {
-        alert("填写不能为空！");
+      else{
+          alert("填写不能为空！");
       }
-    },
+  },
     changePwd() {
-      if (this.form.phone != "" && this.form.password != "") {
+      if (this.form.email != "" && this.form.password != "") {
         axios({
           method: "post",
-          url: "http://8.130.48.144/forget/setPwd",
+          url: "manager/resetPwd",
           data: {
             uid: this.userId,
             old_pwd: this.form.password,
@@ -161,45 +199,58 @@ export default {
         });
       }
     },
-    confirmChange() {
+    login() {
+      localStorage.setItem('token',"wutoken");
+      var a=localStorage.getItem('token');
+      console.log(a);
       if (this.form.email != "" && this.form.password != "") {
-        axios({
-          method: "post",
-          url: "manager/forget/setPwd",
-          data: {
+        axios.post("manager/Login",
+          {
             acc: this.form.email,
-            phoneNumber: this.form.phone,
             pwd: this.form.password,
-            key: this.form.code,
           },
+        ).then((res) => {
+          console.log(axios.defaults.baseURL)
+          let user={};
+          user.Mail=this.form.email;
+          user.UID=res.data.uid;
+          user.hash_code=res.data.hash_code;
+          console.log(user.hash_code);
+          //TODO:页面跳转
+          if(res.data.error_code==0){
+            localStorage.setItem('uid',res.data.uid);
+            localStorage.setItem('Mail',this.form.email);
+            localStorage.setItem('isSuperUser',res.data.is_superUser);
+            localStorage.setItem('token',res.data.hash_code);
+            if(!res.data.is_superUser) {
+              this.$router.push('/settools');
+              this.$store.commit("changeLogin",user);
+            }
+            else{
+              this.$router.push('/borrowProcessing');
+              this.$store.commit("changeLogin",user);
+            }
+          } else if(res.data.error_code==1){
+            alert('存在输入为空！')
+          }
+          else if(res.data.error_code==2){
+            alert('用户不存在！')
+          }
+          else if(res.data.error_code==3){
+            alert('密码错误！请重新输入。')
+          }
+          else alert('请重新输入！')
         })
-          .then((res) => {
-            console.log(res.data.error_code);
-            //TODO:页面跳转
-            if (res.data.error_code == 0) {
-              alert('重置密码成功！')
-              this.$router.push('/teacherLogin');
-            }
-            else if (res.data.error_code == 1) {
-              alert('存在输入为空！')
-            }
-            else if (res.data.error_code == 2) {
-              alert('该邮箱不存在！')
-            }
-            else if (res.data.error_code == 3) {
-              alert('密码错误！请重新输入。')
-            }
-            else alert('验证码错误，请重新输入。')
-          })
       } else {
         alert("填写不能为空");
       }
     },
     register() {
-      console.log(this.form.phone + " : " + this.form.tid);
+      console.log(this.form.phone+" : "+this.form.tid);
       if (this.form.password != "" && this.form.email != "") {
-        if (this.form.password != this.form.password2) alert("两次密码输入不一致。");
-        else {
+        if(this.form.password != this.form.password2) 
+          alert("两次密码输入不一致。");
+        else{
           axios({
             method: "post",
             url: "manager/RegisterTeacherAcc",
@@ -211,17 +262,21 @@ export default {
               key: this.form.code,
               phoneNumber: this.form.phone,
             },
+          }).then((res) => {
+            console.log(res);
+            if(res.data.error_code == 0) {
+              alert('您的注册请求已提交！请等待管理员审核。')
+              this.changeType();
+            }
+            if(res.data.error_code == 1){
+              alert('输入有误！请检查并重新输入')
+            }
+            if(res.data.error_code == 2){
+              alert('该邮箱已注册！')
+            }
           })
-            .then((res) => {
-              console.log(res);
-              if (res.data.error_code == 0) {
-                alert('您的注册请求已提交！请等待管理员审核。')
-                this.$router.push('/teacherLogin');
-              }
-            })
         }
-      }
-      else {
+      } else {
         alert("填写不能为空！");
       }
     },
@@ -235,9 +290,8 @@ export default {
   height: 100vh;
   box-sizing: border-box;
   background-image: url(/MainMenu.jpeg);
-  background-size: 1920px 980px;
+	background-size: 1920px 980px;
 }
-
 .contain {
   width: 60%;
   height: 60%;
@@ -249,17 +303,15 @@ export default {
   border-radius: 20px;
   box-shadow: 0 0 3px #f0f0f0, 0 0 6px #f0f0f0;
 }
-
 .big-box {
   width: 70%;
-  height: 120%;
+  height: 100%;
   position: absolute;
   top: 0;
   left: 30%;
   transform: translateX(0%);
   transition: all 1s;
 }
-
 .big-contain {
   margin-top: 8%;
   width: 100%;
@@ -269,13 +321,11 @@ export default {
   justify-content: center;
   align-items: center;
 }
-
 .btitle {
   font-size: 1.5em;
   font-weight: bold;
   color: rgb(221, 111, 8);
 }
-
 .bform {
   width: 100%;
   height: 40%;
@@ -285,10 +335,9 @@ export default {
   justify-content: space-around;
   align-items: center;
 }
-
 .bform button {
   width: 20%;
-  height: 40px;
+  height: 32px;
   align-items: center;
   border-radius: 9px;
   border: 1px solid rgb(207, 186, 141);
@@ -298,7 +347,6 @@ export default {
   font-size: 0.9em;
   cursor: pointer;
 }
-
 .bform .errTips {
   display: block;
   width: 50%;
@@ -307,7 +355,6 @@ export default {
   font-size: 0.7em;
   margin-left: 1em;
 }
-
 .bform input {
   width: 50%;
   height: 30px;
@@ -317,7 +364,6 @@ export default {
   padding-left: 2em;
   background-color: #f0f0f0;
 }
-
 .bbutton {
   width: 20%;
   height: 40px;
@@ -329,7 +375,6 @@ export default {
   font-size: 0.9em;
   cursor: pointer;
 }
-
 .small-box {
   width: 30%;
   height: 100%;
@@ -342,7 +387,6 @@ export default {
   border-top-left-radius: inherit;
   border-bottom-left-radius: inherit;
 }
-
 .small-contain {
   width: 100%;
   height: 100%;
@@ -351,13 +395,11 @@ export default {
   justify-content: center;
   align-items: center;
 }
-
 .stitle {
   font-size: 1.5em;
   font-weight: bold;
   color: #fff;
 }
-
 .scontent {
   font-size: 0.8em;
   color: #fff;
@@ -365,7 +407,6 @@ export default {
   padding: 2em 4em;
   line-height: 1.7em;
 }
-
 .sbutton {
   width: 60%;
   height: 40px;
@@ -382,7 +423,6 @@ export default {
   left: 0;
   transition: all 0.5s;
 }
-
 .small-box.active {
   left: 100%;
   border-top-left-radius: 0;
@@ -392,7 +432,6 @@ export default {
   transform: translateX(-100%);
   transition: all 1s;
 }
-
 .header {
   height: 50px;
   display: flex;
